+++ cpp/dolfinx/fem/CMakeLists.txt
@@ -4,7 +4,6 @@ set(HEADERS_fem
   ${CMAKE_CURRENT_SOURCE_DIR}/DirichletBC.h
   ${CMAKE_CURRENT_SOURCE_DIR}/DiscreteOperators.h
   ${CMAKE_CURRENT_SOURCE_DIR}/DofMap.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/DofMapBuilder.h
   ${CMAKE_CURRENT_SOURCE_DIR}/ElementDofLayout.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Expression.h
   ${CMAKE_CURRENT_SOURCE_DIR}/FiniteElement.h
@@ -17,6 +16,7 @@ set(HEADERS_fem
   ${CMAKE_CURRENT_SOURCE_DIR}/assemble_matrix_impl.h
   ${CMAKE_CURRENT_SOURCE_DIR}/assemble_scalar_impl.h
   ${CMAKE_CURRENT_SOURCE_DIR}/assemble_vector_impl.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/dofmapbuilder.h
   ${CMAKE_CURRENT_SOURCE_DIR}/dofs_permutation.h
   ${CMAKE_CURRENT_SOURCE_DIR}/dolfin_fem.h
   ${CMAKE_CURRENT_SOURCE_DIR}/evaluate.h
@@ -30,12 +30,12 @@ target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/CoordinateElement.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/DiscreteOperators.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/DofMap.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/DofMapBuilder.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/ElementDofLayout.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/FiniteElement.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/FunctionSpace.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/ReferenceCellGeometry.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/SparsityPatternBuilder.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/dofmapbuilder.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/dofs_permutation.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/petsc.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/utils.cpp
